requirements.txt
pandas
scikit-learn
jupyterlab
torch
torchvision
matplotlib
lightning
tensorboard
poetry
mindsdb_sdk
SpeechRecognition
pdfminer.six
PyPDF2
llama-index-core
llama-parse
llama-index-readers-file
python-dotenv
sentencepiece
accelerate
openai
together